feat: Python側のエラーを細分化
