+++ infer/crates/kanalizer-py/python/kanalizer/__init__.py
@@ -1,6 +1,6 @@
 from typing import TYPE_CHECKING
 from ._rust import __version__, convert, INPUT_CHARS, OUTPUT_CHARS
-from ._error import IncompleteConversionError
+from ._error import IncompleteConversionError, InvalidInputError, EmptyInputError, InvalidCharsError
 
 if TYPE_CHECKING:
     from ._rust import Strategy
@@ -12,4 +12,7 @@ __all__ = [
     "Strategy",
     "convert",
     "IncompleteConversionError",
+    "InvalidInputError",
+    "EmptyInputError",
+    "InvalidCharsError",
 ]
